chore: Add `generateMigrationScript` task to migration sample project
